Android 17.0.0 release 1

* tag 'android-17.0.0_r1': (15196 commits)
  Fix ActiveServicesTest compile failure by adding back import files
  Make ContextLogger do all of its work on the executor
  Fix privileged permission allowlist lookup in PackageManagerShellCommand
  Grant ACCESS_LOCAL_NETWORK and INTERNET permissions to AID_RADIO.
  Grant ACCESS_LOCAL_NETWORK to AID_MEDIA
  MemoryLimiter must ignore EINTR poll results
  Update appwidgetservice to not skip re-loading widgets if intent receivers were not fetched the first time due to error from package manager
  Add additional logging for when widgets are loaded.
  Fix shell access permissions for allowlists.
  Telephony: Support carrier override for network busy message
  Resolve PCC uid in ServiceRecord#postNotification and cancelNotification to avoid FGS failure
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  Enforce suspend restrictions in startOp and noteOp paths
  Block adding toast windows to non-empty tokens.
  Skip serialization of huge vibration effects in NotificationChannel
  ...

+++ core/java/com/android/internal/protolog/LogcatOnlyProtoLogImpl.java
@@ -45,7 +45,7 @@ public class LogcatOnlyProtoLogImpl implements IProtoLog {
 
     @Override
     public void log(@NonNull LogLevel logLevel, @NonNull IProtoLogGroup group, long messageHash,
-            int paramsMask, @Nullable Object[] args) {
+            long paramsMask, @Nullable Object[] args) {
         throw new UnsupportedOperationException("Not supported when using LogcatOnlyProtoLogImpl");
     }
